Enable ED25519 E2E tests
OpenSSL finally caught up.

Closes: #11
